JSON reader writer connected with manageHouse controller. Queue and JSONQueue files were removed.
