+++ src/views/houseAdd_v.java
@@ -1,12 +1,7 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 /**
- *
- * @author Anton Perera
+ * @author: Anton Perera
  */
+
 package views;
 import java.util.*;
 import javax.swing.JOptionPane;
